requirements.txt
pywebview>=5.1
cryptography>=42.0.0
argon2-cffi>=23.1.0
python-vlc>=3.0.20123
mutagen>=1.47.0
platformdirs>=4.2.0
